Move header related tests to Nav.test.tsx
