+++ components/Layout/__tests__/Nav.test.tsx
@@ -1,10 +1,17 @@
 import { render, screen } from "@testing-library/react";
+import userEvent from "@testing-library/user-event";
 import { Nav } from "components/Layout";
 
 describe("Nav component", () => {
-  it("contains navigation accessiblity role", () => {
-    render(<Nav showSidebar={true} toggleSidebar={jest.fn} />);
-    const nav = screen.getByRole("navigation");
-    expect(nav).toBeInTheDocument();
+  it("calls toggle sidebar function when clicking sidebar toggle button", async () => {
+    const toggleSidebar = jest.fn();
+    render(<Nav toggleSidebar={toggleSidebar} showSidebar={false} />);
+
+    const toggleButton = screen.getByRole("button", {
+      name: /toggle sidebar/i,
+    });
+    await userEvent.click(toggleButton);
+
+    expect(toggleSidebar).toBeCalledTimes(1);
   });
 });
